% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertOrdinalToNumeric.R
\name{ConvertOrdinalToNumeric}
\alias{ConvertOrdinalToNumeric}
\title{ConvertOrdinalToNumeric}
\usage{
ConvertOrdinalToNumeric(
  data,
  variables = NULL,
  Data = lifecycle::deprecated(),
  Variables = lifecycle::deprecated()
)
}
\arguments{
\item{data}{The dataframe containing the variables.}

\item{variables}{A character vector specifying the names of variables to consider. If NULL, all columns of the dataframe will be considered.}

\item{Data}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}

\item{Variables}{\strong{Deprecated} (since 19.15.0). Use \code{variables} instead.}
}
\value{
The dataframe with ordinal variables potentially converted to numeric.
}
\description{
Convert ordinal variables in a dataframe to numeric if they contain numeric values in their character representation.
}
