% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFacetLabels.R
\name{createFacetLabels}
\alias{createFacetLabels}
\title{Create facet labels for ggplot2 based on variable labels in a data frame}
\usage{
createFacetLabels(data, DataFrame = lifecycle::deprecated())
}
\arguments{
\item{data}{A data frame containing variable labels.}

\item{DataFrame}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}
}
\value{
A character vector containing facet labels.
}
\description{
This function takes a data frame containing variable labels and creates facet labels
suitable for use with ggplot2 facet functions.
}
