% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateProjectFolders.R
\name{CreateProjectFolders}
\alias{CreateProjectFolders}
\title{Create Project Folder Structure}
\usage{
CreateProjectFolders(base_path = ".")
}
\arguments{
\item{base_path}{A character string specifying the base directory where the folders should be created. Defaults to the current working directory.}
}
\value{
A message indicating that the project structure has been created successfully.
}
\description{
This function creates a project folder structure with the following directories:
\itemize{
\item Data/
\itemize{
\item Raw/
\item Clean/
}
\item Scripts/
\item Reports/
}
}
