% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBinaryVars.R
\name{getBinaryVars}
\alias{getBinaryVars}
\title{Identify Binary Variables}
\usage{
getBinaryVars(DataFrame, Ordinal = TRUE, Revalued = TRUE)
}
\arguments{
\item{DataFrame}{A dataframe to analyze for binary variables.}

\item{Ordinal}{Logical. If TRUE, ordinal factors are included in the search
for binary variables. Default is TRUE.}

\item{Revalued}{Logical. If TRUE, the function checks factors and their levels;
otherwise, it checks for variables with two unique values.}
}
\value{
A character vector containing the names of binary variables.
}
\description{
This function identifies and returns a list of binary variables
in a dataframe. Binary variables are defined as having exactly two unique values
or levels. The function supports options for handling ordinal factors and
revalued data.
}
