% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCatVars.R
\name{getCatVars}
\alias{getCatVars}
\title{Get Categorical Variables}
\usage{
getCatVars(data, Ordinal = TRUE, DataFrame = lifecycle::deprecated())
}
\arguments{
\item{data}{The data frame from which to extract categorical variables.}

\item{Ordinal}{Logical, indicating whether to include ordinal variables.}

\item{DataFrame}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}
}
\value{
A character vector containing the names of categorical variables.
}
\description{
Extracts categorical variables from a data frame.
}
