% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumVars.R
\name{getNumVars}
\alias{getNumVars}
\title{Get Numeric Variables}
\usage{
getNumVars(data, Ordinal = FALSE, DataFrame = lifecycle::deprecated())
}
\arguments{
\item{data}{The data frame from which to extract numeric variables.}

\item{Ordinal}{Logical, indicating whether to include ordinal variables.}

\item{DataFrame}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}
}
\value{
A character vector containing the names of numeric variables.
}
\description{
Extracts numeric variables from a data frame.
}
