% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertValues.R
\name{InsertValues}
\alias{InsertValues}
\title{Insert Values into a String Array}
\usage{
InsertValues(vec, values, after_value, location = "after")
}
\arguments{
\item{vec}{A character vector in which values will be inserted.}

\item{values}{A character vector or a single string to insert into the original vector.}

\item{after_value}{A character string indicating the element in \code{vec} before or after which the values will be inserted. Should have a length of one}

\item{location}{A character string specifying whether to insert \code{values} "before" or "after" the \code{after_value}. Defaults to "after".}
}
\value{
A character vector with the values inserted.
}
\description{
This function inserts one or more strings into an array of strings at specified locations.
}
