% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTable1.R
\name{MakeTable1}
\alias{MakeTable1}
\title{Create Summary Table using gtsummary}
\usage{
MakeTable1(
  data,
  variables = NULL,
  TreatOrdinalAs = "Continuous",
  AutoDetectDistribution = FALSE,
  IncludeMissing = "ifany",
  DataFrame = lifecycle::deprecated(),
  Variables = lifecycle::deprecated()
)
}
\arguments{
\item{data}{The dataframe to create the summary table from.}

\item{variables}{Optional. A character vector specifying the names of variables to include in the summary table. If NULL, all variables are included.}

\item{TreatOrdinalAs}{Character. Specifies how ordinal variables should be treated. Can be "Continuous", "Categorical", or "Both".}

\item{AutoDetectDistribution}{Logical. If TRUE, the function will attempt to automatically detect the distribution of variables. Default is FALSE.}

\item{IncludeMissing}{Character matching gtsummary criteria. Can be "no", "ifany", or "always". Default is "ifany"}

\item{DataFrame}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}

\item{Variables}{\strong{Deprecated} (since 19.15.0). Use \code{variables} instead.}
}
\value{
A summary table created using gtsummary.
}
\description{
This function is a wrapper around \code{gtsummary::tbl_summary} that ensures continuous variables are treated as continuous.
}
