% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureWGCNA.R
\name{PlotFeatureModuleTraitHeatmap}
\alias{PlotFeatureModuleTraitHeatmap}
\title{Plot feature module-trait associations}
\usage{
PlotFeatureModuleTraitHeatmap(module_trait_obj)
}
\arguments{
\item{module_trait_obj}{A \code{feature_module_trait_obj}.}
}
\value{
A ggplot heatmap.
}
\description{
Plot feature module-trait associations
}
