% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureWGCNA.R
\name{PlotFeatureWGCNADendrogram}
\alias{PlotFeatureWGCNADendrogram}
\title{Plot a feature WGCNA dendrogram}
\usage{
PlotFeatureWGCNADendrogram(wgcna_obj)
}
\arguments{
\item{wgcna_obj}{A \code{feature_wgcna_obj}.}
}
\value{
Invisibly returns \code{wgcna_obj} after drawing base graphics.
}
\description{
Plot a feature WGCNA dendrogram
}
