% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForestFromTable.R
\name{plotForestFromTable}
\alias{plotForestFromTable}
\title{Create a Forest Plot from Univariate Regression Tables}
\usage{
plotForestFromTable(UnivariateRegressionTables, pSize = 2, Flip = FALSE)
}
\arguments{
\item{UnivariateRegressionTables}{A list containing formatted regression tables and styling information. Expected structure:
\itemize{
\item \code{FormattedTable$tbls}: A list of tables, each containing a \code{table_body} dataframe.
\item \code{LargeTable$table_styling$header}: A dataframe with a \code{label} and \code{spanning_header} column for headers.
}}

\item{pSize}{Numeric. Size of the points in the plot. Default is 2.}

\item{Flip}{Logical. If \code{FALSE}, outcomes are facets and predictors/terms
are rows. If \code{TRUE}, predictors/terms are facets and outcomes are rows.}
}
\value{
A ggplot object representing the forest plot.
}
\description{
This function generates a forest plot from a list of formatted univariate regression tables.
}
\examples{
# Example usage:
# plotForestFromTable(UnivariateRegressionTables)
}
