% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPartialRegressionScatter.R
\name{PlotPartialRegressionScatter}
\alias{PlotPartialRegressionScatter}
\title{Partial Regression Plot}
\usage{
PlotPartialRegressionScatter(
  data,
  IndepVar,
  DepVar,
  covariates = NULL,
  Relabel = TRUE,
  DataFrame = lifecycle::deprecated(),
  Covariates = lifecycle::deprecated()
)
}
\arguments{
\item{data}{The dataset to use.}

\item{IndepVar}{A string specifying the independent variable.}

\item{DepVar}{A string specifying the dependent variable.}

\item{covariates}{A character vector of covariate names for adjustment. Defaults to NULL.}

\item{Relabel}{Logical indicating whether to use labelled names from the data (using sjlabelled::get_label). Defaults to TRUE.}

\item{DataFrame}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}

\item{Covariates}{\strong{Deprecated} (since 19.15.0). Use \code{covariates} instead.}
}
\value{
A list containing:
\item{plot}{A ggplot2 object representing the partial regression plot.}
\item{method}{The correlation method (as provided).}
\item{Relabel}{Logical; whether relabeling was applied.}
\item{Covariates}{The vector of covariates.}
\item{r2}{The R-squared of the partial regression model.}
\item{p_value}{The p-value for the independent variable coefficient.}
\item{n}{The sample size (number of complete cases).}
\item{equation}{The regression equation string.}
}
\description{
Generate a partial regression plot for a specified independent and dependent variable
while adjusting for covariates. In addition to the figure, key parameters such as
the correlation method, whether relabeling was used, the covariates, R-squared, p-value, and sample size
are returned.
}
\examples{
data(SampleData)
data(SampleVariableTypes)

Labelled <- RevalueData(SampleData, SampleVariableTypes)$RevaluedData

result <- PlotPartialRegressionScatter(
  Labelled,
  IndepVar = "age",
  DepVar = "AXL",
  covariates = "Adiponectin"
)

result$plot
}
