% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTimeDistrubition.R
\name{PlotTimeDistribution}
\alias{PlotTimeDistribution}
\title{Plot Time Distribution}
\usage{
PlotTimeDistribution(
  data,
  DateVariable = "Date",
  Data = lifecycle::deprecated()
)
}
\arguments{
\item{data}{The data frame containing the time-based data.}

\item{DateVariable}{The name of the column in the data frame containing the date information. Default is "Date".}

\item{Data}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}
}
\value{
A ggplot object displaying the distribution of time-based data.
}
\description{
This function plots the distribution of time-based data.
}
