% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeString.R
\name{removeString}
\alias{removeString}
\title{Remove Strings from a Vector}
\usage{
removeString(Orig, Remove)
}
\arguments{
\item{Orig}{The original vector containing strings.}

\item{Remove}{The vector of strings to be removed from the original vector.}
}
\value{
A vector containing the strings from the original vector that were not present in the removal vector.
}
\description{
This function removes strings from a vector that are present in another vector.
}
