% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceMissingCode.R
\name{ReplaceMissingCode}
\alias{ReplaceMissingCode}
\title{Replace Missing Codes with NA}
\usage{
ReplaceMissingCode(
  data,
  codebook,
  DataFrame = lifecycle::deprecated(),
  VariableCodebook = lifecycle::deprecated()
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{codebook}{A data frame containing the variable codebook. It should have columns \code{Variable} and \code{MissingCode}, where \code{Variable} specifies the variable name in the data frame and \code{MissingCode} specifies the code to be replaced with \code{NA}.}

\item{DataFrame}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}

\item{VariableCodebook}{\strong{Deprecated} (since 19.15.0). Use \code{codebook} instead.}
}
\value{
A data frame with specified missing codes replaced by \code{NA}.
}
\description{
This function replaces specified missing codes in a data frame with \code{NA} values based on a given variable codebook.
}
