% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceMissingLabels.R
\name{ReplaceMissingLabels}
\alias{ReplaceMissingLabels}
\title{Replace Missing Labels in Dataframe Columns}
\usage{
ReplaceMissingLabels(data, df = lifecycle::deprecated())
}
\arguments{
\item{data}{A dataframe.}

\item{df}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}
}
\value{
The input dataframe with missing labels replaced.
}
\description{
This function iterates through the columns of a dataframe and assigns the column name as the label to any column that does not have a label.
}
