% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReValueFactors.R
\name{ReValueFactors}
\alias{ReValueFactors}
\title{Revalue Factors}
\usage{
ReValueFactors(
  data,
  codebook,
  DatatoRevalue = lifecycle::deprecated(),
  VarTypes = lifecycle::deprecated()
)
}
\arguments{
\item{data}{The dataset to be revalued.}

\item{codebook}{A data frame containing information about the variables and how they should be revalued.
It should have columns: Variable (variable names), Recode (yes/no for recoding),
and Code (the revalue codes separated by "=" and ","). DO NOT USE COMMAS ANYWHERE ELSE IN THIS COLUMN.}

\item{DatatoRevalue}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}

\item{VarTypes}{\strong{Deprecated} (since 19.15.0). Use \code{codebook} instead.}
}
\value{
The revalued dataset.
}
\description{
This function revalues factor variables in a dataset according to the specifications provided in a codebook.
}
