% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseFactorLevels.R
\name{reverseFactorLevels}
\alias{reverseFactorLevels}
\title{Reverse Levels of Categorical Factors}
\usage{
reverseFactorLevels(data, variables, df = lifecycle::deprecated())
}
\arguments{
\item{data}{A dataframe containing the categorical variables to be reversed.}

\item{variables}{A character vector of column names in the dataframe to reverse levels.
These columns must be categorical factors.}

\item{df}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}
}
\value{
A dataframe with the levels of the specified factors reversed.
Columns not specified in \code{variables} remain unchanged.
}
\description{
This function reverses the levels of specified categorical variables in a given dataframe.
}
