% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleData.R
\docType{data}
\name{SampleData}
\alias{SampleData}
\title{SampleData for practicing SciDataReportR functions}
\format{
\subsection{\code{SampleData}}{

A data frame with 333 rows and 131 columns:
\describe{
\item{Diagnosis}{Control or impaired}
}
}
}
\source{
\url{http://appliedpredictivemodeling.com/data}
}
\usage{
SampleData
}
\description{
A sample dataset adapted from the AppliedPredictive Modeling R package with additional simulated variables, noise and missingness added. The original dataset Craig-Schapiro et al. (2011) describes a clinical study of 333 patients where
laboratory measurements are used to predict which subjects are most likely to develop cognitive impairment, such as Alzheimer's disease.
}
\keyword{datasets}
