% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciDataReportR-package.R
\docType{package}
\name{SciDataReportR-package}
\alias{SciDataReportR}
\alias{SciDataReportR-package}
\title{SciDataReportR: Workflow Infrastructure for Reproducible Scientific Data Reporting}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Workflow infrastructure for turning labelled, researcher-native life science data frames into reproducible exploratory analyses, metadata-aware reports, statistical comparisons, dimensionality reduction workflows, clustering/projection pipelines, scientific visualizations, and report-ready outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RDastgh1/SciDataReportR}
  \item \url{https://rdastgh1.github.io/SciDataReportR/}
  \item Report bugs at \url{https://github.com/RDastgh1/SciDataReportR/issues}
}

}
\author{
\strong{Maintainer}: Raha Dastgheyb \email{raha@jhmi.edu} (\href{https://orcid.org/0000-0002-4992-711X}{ORCID})

Authors:
\itemize{
  \item Raha Dastgheyb \email{raha@jhmi.edu} (\href{https://orcid.org/0000-0002-4992-711X}{ORCID})
}

}
\keyword{internal}
