% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateDataDictionary.R
\name{UpdateDataDictionary}
\alias{UpdateDataDictionary}
\title{Update an existing data dictionary with new variables and types}
\usage{
UpdateDataDictionary(OldDataDictionary, NewDataFrame)
}
\arguments{
\item{OldDataDictionary}{The existing data dictionary to be updated.}

\item{NewDataFrame}{The new data frame containing additional variables.}
}
\value{
A list containing the new variables and the updated data dictionary.
}
\description{
This function updates an existing data dictionary with new variables and their corresponding types.
}
