% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddToCodebook.R
\name{AddToCodebook}
\alias{AddToCodebook}
\title{Add a new variable to a codebook}
\usage{
AddToCodebook(
  codebook,
  VariableName,
  VariableLabel = NA,
  VariableType = NA,
  VariableCategory = NA,
  VariableRecode = NA,
  VariableCode = NA,
  VariableExclude = NA,
  VariableNotes = NA,
  CB = lifecycle::deprecated()
)
}
\arguments{
\item{codebook}{A data frame representing the codebook.}

\item{VariableName}{The name of the variable.}

\item{VariableLabel}{The label of the variable (default: same as VariableName).}

\item{VariableType}{The type of the variable (e.g., numeric, categorical).}

\item{VariableCategory}{The category to which the variable belongs.}

\item{VariableRecode}{The recode information for the variable.}

\item{VariableCode}{The code associated with the variable.}

\item{VariableExclude}{A flag indicating whether the variable should be excluded.}

\item{VariableNotes}{Any additional notes or comments about the variable.}

\item{CB}{\strong{Deprecated} (since 19.15.0). Use \code{codebook} instead.}
}
\value{
A data frame representing the updated codebook with the new variable added.
}
\description{
This function adds a new variable entry to an existing codebook.
}
\examples{
# Create an empty codebook
codebook <- data.frame(Variable = character(0), Label = character(0),
                       Type = character(0), Category = character(0),
                       Recode = character(0), Code = character(0),
                       Exclude = logical(0), Notes = character(0))

# Add a new variable to the codebook
codebook <- AddToCodebook(codebook, "Age", "Age of participants", "numeric", "Demographics")

}
