% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertOrdinalToNumeric.R
\name{ConvertOrdinalToNumeric}
\alias{ConvertOrdinalToNumeric}
\title{ConvertOrdinalToNumeric}
\usage{
ConvertOrdinalToNumeric(Data, Variables = NULL)
}
\arguments{
\item{Data}{The dataframe containing the variables.}

\item{Variables}{A character vector specifying the names of variables to consider. If NULL, all columns of the dataframe will be considered.}
}
\value{
The dataframe with ordinal variables potentially converted to numeric.
}
\description{
Convert ordinal variables in a dataframe to numeric if they contain numeric values in their character representation.
}
