% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePCATable.R
\name{CreatePCATable}
\alias{CreatePCATable}
\title{Create PCA Table and Visualization}
\usage{
CreatePCATable(
  Data,
  VarsToReduce,
  VariableCategories = NULL,
  minThresh = 0.85,
  scale = TRUE,
  center = TRUE,
  Relabel = TRUE,
  Ordinal = FALSE,
  numComponents = NULL
)
}
\arguments{
\item{Data}{The dataset containing the variables for PCA.}

\item{VarsToReduce}{A character vector specifying the variables to include in the PCA.}

\item{VariableCategories, }{categorical, annotates the lollipop plot with color}

\item{minThresh}{The minimum threshold for cumulative proportion of variance (default is 0.85).}

\item{scale}{Logical, indicating whether to scale the data (default is TRUE).}

\item{center}{Logical, indicating whether to center the data (default is TRUE).}

\item{Relabel}{Optional argument to relabel the components.}

\item{Ordinal}{Logical, indicating whether ordinal variables should be handled.}

\item{numComponents}{Number of principal components to compute.}
}
\value{
A list containing PCA results and visualizations.
}
\description{
Perform principal component analysis (PCA) on specified variables and create visualizations.
}
