% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateStatisticsTable.R
\name{CreateStatisticsTable}
\alias{CreateStatisticsTable}
\title{Create Statistics Table}
\usage{
CreateStatisticsTable(Data, TargetVar)
}
\arguments{
\item{Data}{The data frame containing the variables of interest.}

\item{TargetVar}{The target variable for which statistics will be calculated.}
}
\value{
A formatted HTML table displaying statistics.
}
\description{
Generate a table of statistics including means, standard deviations, counts, and p-values.
}
