% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSummaryTable.R
\name{CreateSummaryTable}
\alias{CreateSummaryTable}
\title{Create Summary Table}
\usage{
CreateSummaryTable(Data, Variables, numdecimals = 2, Relabel = TRUE)
}
\arguments{
\item{Data}{The dataset containing the variables of interest.}

\item{Variables}{A character vector specifying the variables for which summary statistics will be calculated.}

\item{numdecimals}{Number of decimal places to round the summary statistics.}

\item{Relabel}{Logical, indicating whether to use variable labels as column headers.}
}
\value{
A formatted HTML table displaying summary statistics.
}
\description{
Generate a descriptive summary table for specified variables in a dataset.
}
