% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVariableTypesTemplate.R
\name{CreateVariableTypesTemplate}
\alias{CreateVariableTypesTemplate}
\title{Create a Template for Variable Types}
\usage{
CreateVariableTypesTemplate(
  data,
  CSVFileName = NULL,
  GuessCategorical = TRUE,
  DataFrame = lifecycle::deprecated()
)
}
\arguments{
\item{data}{A data frame containing the variables to be summarized.}

\item{CSVFileName}{A string specifying the path and name of the CSV file to save the summary.
If NULL (the default), the CSV file will not be created.}

\item{GuessCategorical}{A logical variable specifying if the function should guess what variables are categorical based on having <= 5 unique values}

\item{DataFrame}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}
}
\value{
A data frame with the following columns:
\describe{
\item{Variable}{The names of the variables in the input data frame.}
\item{Label}{The labels of the variables, if available; otherwise, the variable names.}
\item{Type}{The data types of the variables, converted to more user-friendly descriptions.}
\item{Category}{A placeholder column for categorizing variables (default is NA).}
\item{Recode}{A placeholder column for recoding information (default is NA).}
\item{Code}{A placeholder column for code information (default is NA).}
\item{Notes}{A placeholder column for any additional notes (default is an empty string).}
\item{Exclude}{A placeholder column for exclusion flags (default is NA).}
}
}
\description{
This function generates a data frame that summarizes the types and labels of the variables in a given data frame.
It optionally saves this summary to a CSV file.
}
\examples{
df <- data.frame(
  num = c(1.1, 2.2),
  int = c(1L, 2L),
  fact = factor(c("A", "B")),
  char = c("a", "b"),
  date = as.Date(c("2021-01-01", "2021-01-02"))
)
CreateVariableTypesTemplate(df)
CreateVariableTypesTemplate(df, file.path(tempdir(), "variable_types.csv"))

}
