% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateZScorePlot.R
\name{CreateZScorePlot}
\alias{CreateZScorePlot}
\title{Create a Z-Score Plot with Statistical Significance}
\usage{
CreateZScorePlot(
  Data,
  TargetVar,
  Variables,
  VariableCategories = NULL,
  Relabel = TRUE,
  sort = TRUE,
  RemoveXAxisLabels = TRUE,
  Ordinal = FALSE
)
}
\arguments{
\item{Data}{A data frame containing the target variable and variables of interest.}

\item{TargetVar}{A string specifying the name of the target variable (grouping variable).}

\item{Variables}{A vector of strings specifying the names of the variables to be plotted.}

\item{VariableCategories}{An optional vector of categories corresponding to the variables.}

\item{Relabel}{A logical value indicating whether to relabel the variables. Default is TRUE.}

\item{sort}{A logical value indicating whether to sort the variables by p-value and category. Default is TRUE.}

\item{RemoveXAxisLabels}{A logical value indicating whether to remove X-axis labels. Default is TRUE.}
}
\value{
A ggplot object representing the Z-score plot.
}
\description{
This function generates a Z-score plot for the provided data, highlighting the mean values, standard errors, and p-values for significance testing between groups.
}
