% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormattedDataDictionary.R
\name{FormattedDataDictionary}
\alias{FormattedDataDictionary}
\title{Create a formatted data dictionary table}
\usage{
FormattedDataDictionary(DataFrame, numdecimals = 2)
}
\arguments{
\item{DataFrame}{The data frame for which the data dictionary is to be created.}

\item{numdecimals}{Number of decimals to display for numeric variables (default: 2).}
}
\value{
A formatted data dictionary table.
}
\description{
This function generates a formatted data dictionary table using the specified data frame.
The table includes variable names, labels, types, and additional formatting based on variable types.
}
