% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_starcaption.R
\name{geom_starcaption}
\alias{geom_starcaption}
\title{Add a Caption Explaining Star Annotations}
\usage{
geom_starcaption()
}
\value{
A \code{labs()} object that can be added to a ggplot, specifically designed for the SciDataReportR heatmaps
}
\description{
This function adds a caption to a ggplot explaining the meaning of star annotations (*, **, ***).
}
