% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumVars.R
\name{getNumVars}
\alias{getNumVars}
\title{Get Numeric Variables}
\usage{
getNumVars(DataFrame, Ordinal = TRUE)
}
\arguments{
\item{DataFrame}{The data frame from which to extract numeric variables.}

\item{Ordinal}{Logical, indicating whether to include ordinal variables.}
}
\value{
A character vector containing the names of numeric variables.
}
\description{
Extracts numeric variables from a data frame.
}
