% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_DataDictionary.R
\name{Make_DataDictionary}
\alias{Make_DataDictionary}
\title{Create a data dictionary for a data frame}
\usage{
Make_DataDictionary(DataFrame, numdecimals = 2)
}
\arguments{
\item{DataFrame}{The data frame for which the data dictionary is to be created.}

\item{numdecimals}{Number of decimals to display for numeric variables (default: 2).}
}
\value{
A data frame representing the data dictionary.
}
\description{
This function generates a data dictionary for a given data frame, including variable names, labels,
types, missing value statistics, summary statistics, and other relevant information.
}
