% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTable1.R
\name{MakeTable1}
\alias{MakeTable1}
\title{Create Summary Table using gtsummary}
\usage{
MakeTable1(
  DataFrame,
  Variables = NULL,
  TreatOrdinalAs = "Continuous",
  AutoDetectDistribution = FALSE,
  IncludeMissing = "ifany"
)
}
\arguments{
\item{DataFrame}{The dataframe to create the summary table from.}

\item{Variables}{Optional. A character vector specifying the names of variables to include in the summary table. If NULL, all variables are included.}

\item{TreatOrdinalAs}{Character. Specifies how ordinal variables should be treated. Can be "Continuous", "Categorical", or "Both".}

\item{AutoDetectDistribution}{Logical. If TRUE, the function will attempt to automatically detect the distribution of variables. Default is FALSE.}

\item{IncludeMissing}{Character matching gtsummary criteria. Can be "no", "ifany", or "always". Default is "ifany"}
}
\value{
A summary table created using gtsummary.
}
\description{
This function is a wrapper around \code{gtsummary::tbl_summary} that ensures continuous variables are treated as continuous.
}
