% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notInOperator.R
\name{not_in_operator}
\alias{not_in_operator}
\alias{\%!in\%}
\alias{\%notin\%}
\title{Negated "in" Operator}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{The element to be tested for absence in the vector.}

\item{y}{The vector in which to search for the element.}
}
\value{
TRUE if the element is not found in the vector, otherwise FALSE.
}
\description{
This custom operator returns TRUE if the element on the left-hand side of the operator is not found
in the vector on the right-hand side when using the \%in\% operator. Otherwise, it returns FALSE.
}
