% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot2GroupStats.R
\name{Plot2GroupStats}
\alias{Plot2GroupStats}
\title{Plot and Summarize Group Statistics}
\usage{
Plot2GroupStats(
  Data,
  Variables,
  VariableCategories = NULL,
  impClust,
  normalClust
)
}
\arguments{
\item{Data}{A data frame containing the data to be analyzed.}

\item{Variables}{A character vector of variable names to be included in the analysis.}

\item{VariableCategories}{A data frame containing variable categories (optional).}

\item{impClust}{A character string representing the important cluster.}

\item{normalClust}{A character string representing the normal cluster.}
}
\value{
A list containing a plot, a summary table, and a p-value table.
}
\description{
This function creates plots and summary tables to compare two groups within a dataset.
}
