% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAnovaRelationshipsMatrix.R
\name{PlotAnovaRelationshipsMatrix}
\alias{PlotAnovaRelationshipsMatrix}
\title{Plot ANOVA Relationships Matrix}
\usage{
PlotAnovaRelationshipsMatrix(
  Data,
  CatVars,
  ContVars,
  Covariates = NULL,
  Relabel = TRUE,
  Parametric = TRUE,
  Ordinal = FALSE,
  min_n = 4,
  eps = 1e-08
)
}
\arguments{
\item{Data}{The data frame containing the variables of interest.}

\item{CatVars}{Character vector of categorical variable names.}

\item{ContVars}{Character vector of continuous variable names.}

\item{Covariates}{Optional character vector of covariate names for ANCOVA analysis.}

\item{Relabel}{Logical indicating whether to relabel variables with their labels (default is TRUE).}

\item{Parametric}{Logical indicating whether to use parametric (ANOVA) or non-parametric (Kruskal-Wallis) tests (default is TRUE).}

\item{Ordinal}{Logical, indicating whether ordinal variables should be considered.}
}
\value{
A list containing three ggplot objects: p (scatter plot without multiple comparison correction), p_FDR (scatter plot with FDR correction), and pvaltable (data frame of p-values and significance).
}
\description{
This function plots the relationship between continuous and categorical variables
using ANOVA or Kruskal-Wallis tests. It generates a "heatmap" with points
colored and shaped based on statistical significance and effect size.
}
