% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAssociations.R
\name{PlotAssociations}
\alias{PlotAssociations}
\title{Plot Associations}
\usage{
PlotAssociations(
  data,
  Var1,
  Var2,
  Ordinal = FALSE,
  DataFrame = lifecycle::deprecated()
)
}
\arguments{
\item{data}{The data frame containing the variables of interest.}

\item{Var1}{The name of the first variable.}

\item{Var2}{The name of the second variable.}

\item{Ordinal}{Logical, indicating whether ordinal variables should be included.}

\item{DataFrame}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}
}
\value{
A ggplot object representing the relationship between the variables.
}
\description{
This function generates scatter plots or box plots to visualize the relationship between two variables.
}
