% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBlandAltman.R
\name{PlotBlandAltman}
\alias{PlotBlandAltman}
\title{Plot Bland-Altman Agreement Plot}
\usage{
PlotBlandAltman(
  data,
  Variable1,
  Variable2,
  DataFrame = lifecycle::deprecated()
)
}
\arguments{
\item{data}{A data frame containing the variables to compare.}

\item{Variable1}{The name of the first variable (as a string) to compare.}

\item{Variable2}{The name of the second variable (as a string) to compare.}

\item{DataFrame}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}
}
\value{
A list containing:
\item{plot}{A ggplot2 object of the Bland-Altman plot.}
\item{stats}{A list of Bland-Altman statistics from the BlandAltmanLeh package.}
}
\description{
Generates a Bland-Altman plot to visualize the agreement between two variables.
}
\note{
This function is adapted from code written by Eran Shorer.
}
