% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCategoricalDistributions.R
\name{PlotCategoricalDistributions}
\alias{PlotCategoricalDistributions}
\title{Plot categorical distributions}
\usage{
PlotCategoricalDistributions(
  data,
  variables = NULL,
  Relabel = TRUE,
  Ordinal = TRUE,
  LabelType = "percent",
  MissingLabel = "Missing",
  DataFrame = lifecycle::deprecated(),
  Variables = lifecycle::deprecated()
)
}
\arguments{
\item{data}{The dataframe containing the variables to be plotted.}

\item{variables}{Optional. A character vector specifying the names of the
categorical variables to be plotted. If NULL, categorical variables are
automatically detected.}

\item{Relabel}{Logical. If TRUE, missing labels in the dataframe are replaced
with column names as labels for plotting.}

\item{Ordinal}{Logical, indicating whether ordinal variables should be included.}

\item{LabelType}{Character. Either "percent" or "count", indicating what
should be shown on the x-axis and inside the bars.}

\item{MissingLabel}{Character label to use for missing values.}

\item{DataFrame}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}

\item{Variables}{\strong{Deprecated} (since 19.15.0). Use \code{variables} instead.}
}
\value{
A ggplot object visualizing the distributions of categorical variables.
}
\description{
This function creates plots to visualize the distributions of categorical
variables in a dataframe.
}
