% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCategoricalDistributions.R
\name{PlotCategoricalDistributions}
\alias{PlotCategoricalDistributions}
\title{Plot Categorical Distributions}
\usage{
PlotCategoricalDistributions(
  DataFrame,
  Variables = NULL,
  Relabel = TRUE,
  Ordinal = TRUE
)
}
\arguments{
\item{DataFrame}{The dataframe containing the variables to be plotted.}

\item{Variables}{Optional. A character vector specifying the names of the categorical variables to be plotted. If NULL, categorical variables are automatically detected.}

\item{Relabel}{Logical. If TRUE, missing labels in the dataframe are replaced with column names as labels for plotting.}
}
\value{
A plot visualizing the distributions of categorical variables.
}
\description{
This function creates plots to visualize the distributions of categorical variables in a dataframe.
}
