% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCatInteractionEffectsMatrix.R
\name{PlotCatInteractionEffectsMatrix}
\alias{PlotCatInteractionEffectsMatrix}
\title{Plot Categorical Interaction Effects Matrix}
\usage{
PlotCatInteractionEffectsMatrix(
  Data,
  xVars,
  yVars = NULL,
  xVarLabels = NULL,
  yVarLabels = NULL,
  interVar
)
}
\arguments{
\item{Data}{The dataset containing the variables.}

\item{xVars}{A character vector of the names of the x-axis categorical variables.}

\item{yVars}{A character vector of the names of the y-axis categorical variables. Defaults to NULL,
in which case it takes the same values as xVars.}

\item{xVarLabels}{A character vector of labels for the x-axis variables. Defaults to NULL,
in which case it takes the same values as xVars.}

\item{yVarLabels}{A character vector of labels for the y-axis variables. Defaults to NULL,
in which case it takes the same values as yVars.}

\item{interVar}{The name of the interaction variable.}
}
\value{
A list containing matrices of interaction coefficients, p-values, ggplot objects for visualizations,
and tables of FDR-corrected p-values.
}
\description{
This function calculates and visualizes the interaction effects between categorical variables.
}
