% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotChiSqCovar.R
\name{PlotChiSqCovar}
\alias{PlotChiSqCovar}
\title{Plot Chi-Square Test with Covariate}
\usage{
PlotChiSqCovar(
  Data,
  xVars,
  yVars,
  covars = NULL,
  Relabel = TRUE,
  Ordinal = TRUE
)
}
\arguments{
\item{Data}{The dataset containing the variables.}

\item{xVars}{A character vector of the names of the x-axis categorical variables.}

\item{yVars}{A character vector of the names of the y-axis categorical variables.}

\item{covars}{A character vector of the names of covariate variables. Defaults to NULL.}

\item{Relabel}{Logical indicating whether to relabel the variables based on their labels in the dataset. Defaults to TRUE.}

\item{Ordinal}{Logical, indicating whether ordinal variables should be included.}
}
\value{
A list containing ggplot objects for visualizations and tables of p-values.
}
\description{
This function conducts chi-square tests with covariates and visualizes the results.
}
