% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotContinuousDistributions.R
\name{PlotContinuousDistributions}
\alias{PlotContinuousDistributions}
\title{Plot Continuous Distributions}
\usage{
PlotContinuousDistributions(
  DataFrame,
  Variables = NULL,
  Relabel = TRUE,
  ncol = 3
)
}
\arguments{
\item{DataFrame}{A data frame containing the variables to be plotted.}

\item{Variables}{A character vector specifying the names of the variables to be plotted. If NULL, all numeric variables in DataFrame will be used.}

\item{Relabel}{Logical, whether to relabel the variables using their labels if available. Default is TRUE.}

\item{ncol}{Number of columns in the facet grid. Default is 3.}
}
\value{
A raincloud plot visualizing the distributions of continuous variables.
}
\description{
This function creates raincloud plots to visualize the distributions of continuous variables in a data frame.
}
