% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCorrelationsHeatmap.R
\name{PlotCorrelationsHeatmap}
\alias{PlotCorrelationsHeatmap}
\title{Plot Correlations Heatmap}
\usage{
PlotCorrelationsHeatmap(
  Data,
  xVars = NULL,
  yVars = NULL,
  covars = NULL,
  method = "pearson",
  Relabel = TRUE,
  Ordinal = FALSE,
  min_n = 3,
  eps = 1e-12
)
}
\arguments{
\item{Data}{The dataset containing the variables.}

\item{xVars}{A character vector of the names of the x-axis variables.}

\item{yVars}{A character vector of the names of the y-axis variables. Defaults to NULL.}

\item{covars}{A character vector of the names of covariate variables. Defaults to NULL.}

\item{method}{The correlation method. Defaults to "pearson".}

\item{Relabel}{A logical indicating whether to relabel variables. Defaults to TRUE.}

\item{Ordinal}{Logical, indicating whether ordinal variables should be included.}

\item{min_n}{Minimum complete cases required for a pairwise correlation.}

\item{eps}{Small tolerance to treat near-constant variables as constant.}
}
\value{
A list containing matrices, ggplot objects for visualizations, and details of the method used.
}
\description{
This function calculates correlations between variables and plots them as a heatmap.
}
