% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDirectionalHeatmaps.R
\name{PlotDirectionalHeatmaps}
\alias{PlotDirectionalHeatmaps}
\title{This function creates directional heatmaps for the associations between continuous and binary variables.
It identifies continuous and binary variables in the dataset and computes correlations using appropriate methods.}
\usage{
PlotDirectionalHeatmaps(
  Data,
  xVars = NULL,
  yVars = NULL,
  Relabel = TRUE,
  Ordinal = TRUE
)
}
\arguments{
\item{Data}{A dataframe containing the data to analyze.}

\item{xVars}{A vector of variable names to be used on the x-axis. If NULL, all continuous and binary variables are included.}

\item{yVars}{A vector of variable names to be used on the y-axis. If NULL, defaults to the same as xVars.}

\item{Relabel}{Logical. If TRUE, variables are relabeled for better readability in the plots.}

\item{Ordinal}{Logical. If TRUE, treats binary variables as ordinal for certain calculations (future functionality).}
}
\value{
A list containing two plots (Unadjusted and FDR Corrected), along with corresponding data.
}
\description{
This function creates directional heatmaps for the associations between continuous and binary variables.
It identifies continuous and binary variables in the dataset and computes correlations using appropriate methods.
}
