% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotInteractionEffectsContinuous.R
\name{PlotInteractionEffectsContinuous}
\alias{PlotInteractionEffectsContinuous}
\title{PlotInteractionEffectsContinuous}
\usage{
PlotInteractionEffectsContinuous(
  Data,
  interVar = NULL,
  xVars = NULL,
  yVars = NULL,
  covars = NULL,
  Relabel = TRUE,
  Ordinal = FALSE
)
}
\arguments{
\item{Data}{The dataset to be analyzed}

\item{interVar}{The variable of interest for the interaction effect}

\item{xVars}{The variables to be plotted on the x-axis}

\item{yVars}{The variables to be plotted on the y-axis}

\item{covars}{Covariates to be included in the analysis}

\item{Relabel}{A logical value indicating whether to relabel the variables with their corresponding labels}

\item{Ordinal}{A logical value indicating whether the variables are ordinal}
}
\value{
A list containing the results of the analysis, including the p-values, significance levels, and the ggplot object.
}
\description{
Plots the interaction effects between continuous variables in a dataset.
}
