% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMissingData.R
\name{PlotMissingData}
\alias{PlotMissingData}
\title{Plot Missing Data}
\usage{
PlotMissingData(DataFrame, Variables = NULL, Relabel = TRUE)
}
\arguments{
\item{DataFrame}{The dataframe containing missing data.}

\item{Variables}{A character vector specifying the variables to be included in the plot. If NULL (default), all variables in the dataframe will be used.}

\item{Relabel}{Logical indicating whether to relabel the columns based on their labels. Default is TRUE.}
}
\value{
A ggplot object visualizing missing data with flipped coordinates.
}
\description{
A wrapper function to plot missing data using the naniar package and flip coordinates.
}
