% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMissingData.R
\name{PlotMissingData}
\alias{PlotMissingData}
\title{Plot Missing Data}
\usage{
PlotMissingData(
  DataFrame,
  Variables = NULL,
  Relabel = TRUE,
  show_perc = TRUE,
  show_perc_var = TRUE,
  cluster = FALSE
)
}
\arguments{
\item{DataFrame}{The dataframe containing missing data.}

\item{Variables}{A character vector specifying the variables to be included in the plot. If NULL (default), all variables in the dataframe will be used.}

\item{Relabel}{Logical indicating whether to relabel the columns based on their labels. Default is TRUE.}

\item{show_perc}{Logical indicating whether to show percentage labels in legend. Default is TRUE.}

\item{show_perc_var}{Logical indicating whether to show percentage of missing for each variable. Default is TRUE.}

\item{cluster}{Logical indicating whether to cluster rows by missingness pattern. Default is FALSE.}
}
\value{
A ggplot object visualizing missing data with variables as rows.
}
\description{
A custom function to plot missing data with variables as rows and observations as columns.
}
