% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNumInteractionEffectsMatrix.R
\name{PlotNumInteractionEffectsMatrix}
\alias{PlotNumInteractionEffectsMatrix}
\title{Plot Numerical Interaction Effects Matrix}
\usage{
PlotNumInteractionEffectsMatrix(
  Data,
  xVars,
  yVars = NULL,
  xVarLabels = NULL,
  yVarLabels = NULL,
  interVar = NULL,
  covars = NULL
)
}
\arguments{
\item{Data}{The dataset containing the variables.}

\item{xVars}{A character vector of the names of the x-axis numerical variables.}

\item{yVars}{A character vector of the names of the y-axis numerical variables. Defaults to NULL.}

\item{xVarLabels}{A character vector of labels for the x-axis variables. Defaults to NULL.}

\item{yVarLabels}{A character vector of labels for the y-axis variables. Defaults to NULL.}

\item{interVar}{The interaction variable. Defaults to NULL.}

\item{covars}{A character vector of the names of covariate variables. Defaults to NULL.}
}
\value{
A list containing matrices, ggplot objects for visualizations, and tables of p-values.
}
\description{
This function calculates interaction effects between numerical variables and plots them as matrices.
}
