% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot 3D PCA Scores}
\usage{
plotPCA(PCAObj, Var = NULL, t = "NULL")
}
\arguments{
\item{PCAObj}{A list containing PCA results, including a \code{Scores} matrix with PCA scores and a \code{CombinedData} dataframe with the original data.}

\item{Var}{An optional string specifying the variable name in \code{CombinedData} to color the points by. Default is \code{NULL}.}

\item{t}{An optional string specifying the type of variable for coloring: \code{"Factor"} for categorical variables or anything else for continuous variables. Default is \code{"NULL"}.}
}
\value{
A Plotly object representing the 3D PCA plot.
}
\description{
This function creates a 3D scatter plot of PCA scores using Plotly. It allows for optional coloring by a variable.
}
