% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPhiHeatmap.R
\name{PlotPhiHeatmap}
\alias{PlotPhiHeatmap}
\title{Plot Phi Heatmap}
\usage{
PlotPhiHeatmap(
  Data,
  xVars = NULL,
  yVars = NULL,
  Relabel = TRUE,
  Ordinal = TRUE
)
}
\arguments{
\item{Data}{A dataframe containing the variables of interest.}

\item{xVars}{A vector of variable names for the x-axis (predictor variables).}

\item{yVars}{A vector of variable names for the y-axis (outcome variables).}

\item{Relabel}{A logical indicating whether to relabel the variables with their labels (default is TRUE).}

\item{Ordinal}{A logical indicating whether to consider ordinal variables when determining binary variables (default is TRUE).}
}
\value{
A list containing the following components:
\item{Unadjusted}{A list of results containing the Phi coefficients, p-values, and number of pairs for each variable combination.}
\item{FDRCorrected}{A list containing FDR-corrected p-values.}
\item{method}{A string indicating that the method used is "Phi".}
\item{Relabel}{The relabel parameter.}
\item{Covariates}{Currently set to NULL, as covariates are not implemented in this function.}
}
\description{
This function generates a heatmap of the Phi coefficient for the association between binary variables
in the provided dataset. The Phi coefficient measures the association between two binary variables,
and the function also performs chi-square tests for each pair of variables.
}
