% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPointCorrelations.R
\name{PlotPointCorrelationsHeatmap}
\alias{PlotPointCorrelationsHeatmap}
\title{Plot Point-Biserial Correlations Between Binary and Continuous Variables}
\usage{
PlotPointCorrelationsHeatmap(
  Data,
  CatVars,
  ContVars,
  Covariates = NULL,
  Relabel = TRUE,
  Ordinal = TRUE
)
}
\arguments{
\item{Data}{A dataframe containing the data.}

\item{CatVars}{A vector of column names representing binary categorical variables.}

\item{ContVars}{A vector of column names representing continuous variables.}

\item{Covariates}{A vector of column names representing covariates (optional).}

\item{Relabel}{A logical value indicating whether to relabel variables using variable labels (default: TRUE).}

\item{Ordinal}{A logical value indicating whether to treat ordinal variables as numeric (default: TRUE).}
}
\value{
A list containing two elements:
\itemize{
\item \code{Unadjusted}: A list with the unadjusted p-value table and corresponding plot.
\item \code{FDRCorrected}: A list with the FDR-adjusted p-value table and corresponding plot.
\item \code{method}: The statistical method used ("R_pb").
\item \code{Relabel}: Logical indicating whether relabeling was applied.
\item \code{Covariates}: The covariates used.
}
}
\description{
This function calculates point-biserial correlations between binary categorical variables and continuous variables,
creates a summary table, and generates heatmap-style plots to visualize the correlations.
}
