% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSigAssociations.R
\name{plotSigAssociations}
\alias{plotSigAssociations}
\title{Plot Significant Associations}
\usage{
plotSigAssociations(
  data,
  AnovaMatrixObject,
  PVar = "p",
  Pthresh = 0.05,
  DataFrame = lifecycle::deprecated()
)
}
\arguments{
\item{data}{The dataset used to generate the scatterplots.}

\item{AnovaMatrixObject}{The output of the PlotAnovaRelationshipsMatrix function.}

\item{PVar}{The name of the column used to filter for significance (default is "P").}

\item{Pthresh}{The significance threshold (default is 0.05).}

\item{DataFrame}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}
}
\value{
A list of scatterplot objects for significant correlations.
}
\description{
Generate ggbetweenstats for significant correlations based on a previously generated anova matrix
}
\examples{
\donttest{
# Build an ANOVA relationships matrix, then plot the significant pairs
av <- PlotAnovaRelationshipsMatrix(
  mtcars,
  CatVars = c("cyl", "gear"),
  ContVars = c("mpg", "wt", "hp")
)

plots <- plotSigAssociations(mtcars, av)

# Display the first significant-association plot
plots[[1]]
}
}
