% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSigCorrelations.R
\name{plotSigCorrelations}
\alias{plotSigCorrelations}
\title{Plot Significant Correlations}
\usage{
plotSigCorrelations(
  data,
  CorrelationHeatmapObject,
  PVar = "P",
  Pthresh = 0.05,
  DataFrame = lifecycle::deprecated()
)
}
\arguments{
\item{data}{The dataset used to generate the scatterplots.}

\item{CorrelationHeatmapObject}{The output of the PlotCorrelationsHeatmap function.}

\item{PVar}{The name of the column used to filter for significance (default is "P").}

\item{Pthresh}{The significance threshold (default is 0.05).}

\item{DataFrame}{\strong{Deprecated} (since 19.15.0). Use \code{data} instead.}
}
\value{
A list of scatterplot objects for significant correlations.
}
\description{
Generate scatterplots for significant correlations based on a previously generated correlation heatmap.
}
\examples{
\donttest{
# Build a correlation heatmap, then plot the significant pairs
ch <- PlotCorrelationsHeatmap(
  mtcars,
  predictor_vars = c("mpg", "wt", "hp"),
  outcome_vars = c("mpg", "wt", "hp")
)

plots <- plotSigCorrelations(mtcars, ch)

# Display the first significant-correlation scatterplot
plots[[1]]
}
}
