% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSigCorrelations.R
\name{plotSigCorrelations}
\alias{plotSigCorrelations}
\title{Plot Significant Correlations}
\usage{
plotSigCorrelations(
  DataFrame,
  CorrelationHeatmapObject,
  PVar = "P",
  Pthresh = 0.05
)
}
\arguments{
\item{DataFrame}{The dataset used to generate the scatterplots.}

\item{CorrelationHeatmapObject}{The output of the PlotCorrelationsHeatmap function.}

\item{PVar}{The name of the column used to filter for significance (default is "P").}

\item{Pthresh}{The significance threshold (default is 0.05).}
}
\value{
A list of scatterplot objects for significant correlations.
}
\description{
Generate scatterplots for significant correlations based on a previously generated correlation heatmap.
}
