% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTimeDistrubition.R
\name{PlotTimeDistribution}
\alias{PlotTimeDistribution}
\title{Plot Time Distribution}
\usage{
PlotTimeDistribution(Data, DateVariable = "Date")
}
\arguments{
\item{Data}{The data frame containing the time-based data.}

\item{DateVariable}{The name of the column in the data frame containing the date information. Default is "Date".}
}
\value{
A ggplot object displaying the distribution of time-based data.
}
\description{
This function plots the distribution of time-based data.
}
