% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceMissingCode.R
\name{ReplaceMissingCode}
\alias{ReplaceMissingCode}
\title{Replace Missing Codes with NA}
\usage{
ReplaceMissingCode(DataFrame, VariableCodebook)
}
\arguments{
\item{DataFrame}{A data frame containing the data.}

\item{VariableCodebook}{A data frame containing the variable codebook. It should have columns \code{Variable} and \code{MissingCode}, where \code{Variable} specifies the variable name in the data frame and \code{MissingCode} specifies the code to be replaced with \code{NA}.}
}
\value{
A data frame with specified missing codes replaced by \code{NA}.
}
\description{
This function replaces specified missing codes in a data frame with \code{NA} values based on a given variable codebook.
}
