% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceMissingLabels.R
\name{ReplaceMissingLabels}
\alias{ReplaceMissingLabels}
\title{Replace Missing Labels in Dataframe Columns}
\usage{
ReplaceMissingLabels(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
The input dataframe with missing labels replaced.
}
\description{
This function iterates through the columns of a dataframe and assigns the column name as the label to any column that does not have a label.
}
