% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevalueData.R
\name{RevalueData}
\alias{RevalueData}
\title{Revalue Data}
\usage{
RevalueData(DatatoRevalue, VarTypes, missingVal = -999, splitchar = ",")
}
\arguments{
\item{DatatoRevalue}{The dataset to be revalued.}

\item{VarTypes}{A data frame containing information about the variables and how they should be revalued.
It should have columns: Variable (variable names), Recode (yes/no for recoding),
Code (the revalue codes separated by ","), Type (Categorical, Double, or Ordinal), and Label (optional label for the variable).}

\item{missingVal}{The value to use for missing values (default is -999).}

\item{splitchar}{The character used to split codes in the Code column (default is ",").}
}
\value{
A list containing the revalued dataset, a warning list, and a list of recoded variables.
}
\description{
This function revalues variables in a dataset according to the specifications provided in a codebook.
}
