% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReValueFactors.R
\name{ReValueFactors}
\alias{ReValueFactors}
\title{Revalue Factors}
\usage{
ReValueFactors(DatatoRevalue, VarTypes)
}
\arguments{
\item{DatatoRevalue}{The dataset to be revalued.}

\item{VarTypes}{A data frame containing information about the variables and how they should be revalued.
It should have columns: Variable (variable names), Recode (yes/no for recoding),
and Code (the revalue codes separated by "=" and ","). DO NOT USE COMMAS ANYWHERE ELSE IN THIS COLUMN.}
}
\value{
The revalued dataset.
}
\description{
This function revalues factor variables in a dataset according to the specifications provided in a codebook.
}
