% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseFactorLevels.R
\name{reverseFactorLevels}
\alias{reverseFactorLevels}
\title{Reverse Levels of Categorical Factors}
\usage{
reverseFactorLevels(df, variables)
}
\arguments{
\item{df}{A dataframe containing the categorical variables to be reversed.}

\item{variables}{A character vector of column names in the dataframe to reverse levels.
These columns must be categorical factors.}
}
\value{
A dataframe with the levels of the specified factors reversed.
Columns not specified in \code{variables} remain unchanged.
}
\description{
This function reverses the levels of specified categorical variables in a given dataframe.
}
