% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciDataReportR-package.R
\docType{package}
\name{SciDataReportR-package}
\alias{SciDataReportR}
\alias{SciDataReportR-package}
\title{SciDataReportR: Package for Generating Scientific Data Reports}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

SciDataReportR is a package designed to facilitate the generation of scientific data reports. It provides functions for creating descriptive summary tables, formatting data for analysis, and enhancing data visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RDastgh1/SciDataReportR}
  \item \url{https://rdastgh1.github.io/SciDataReportR/}
  \item Report bugs at \url{https://github.com/RDastgh1/SciDataReportR/issues}
}

}
\author{
\strong{Maintainer}: Raha Dastgheyb \email{raha@jhmi.edu} (\href{https://orcid.org/signin}{ORCID})

}
\keyword{internal}
