% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateCodebook.R
\name{UpdateCodebook}
\alias{UpdateCodebook}
\title{Update an existing codebook based on a given dataframe}
\usage{
UpdateCodebook(
  Dataframe,
  Codebook,
  RemoveMissing = TRUE,
  ReplaceLabels = FALSE
)
}
\arguments{
\item{Dataframe}{A dataframe for which the codebook needs to be updated.}

\item{Codebook}{A dataframe representing the existing codebook with at least a 'Variable' column.}

\item{RemoveMissing}{Logical; if TRUE, removes variables from the codebook that are not in the dataframe.}

\item{ReplaceLabels}{Logical; if TRUE, replaces outdated labels in the codebook with new ones from the dataframe.}
}
\value{
A list containing:
\itemize{
\item \code{UpdatedCodebook}: The updated codebook dataframe.
\item \code{NewVariables}: Variables present in the dataframe but missing from the original codebook.
\item \code{NotExistingVariables}: Variables in the codebook that are not present in the dataframe.
\item \code{MismatchedLabels}: A dataframe of variables with mismatched labels between the codebook and the dataframe.
}
}
\description{
This function updates a codebook by adding missing variables, removing variables
that no longer exist in the dataframe (if specified), and optionally replacing
outdated labels.
}
