% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_EDATemplate.R, R/use_GetStartedScript.R
\name{use_EDATemplate}
\alias{use_EDATemplate}
\title{Use the EDATemplate Quarto Template}
\usage{
use_EDATemplate(filename = "Scripts/script_GetStarted.R")

use_EDATemplate(filename = "Scripts/script_GetStarted.R")
}
\arguments{
\item{filename}{The name to save the Quarto file as (default: "EDA_Report.qmd").}
}
\description{
Copies the EDATemplate template to the working directory.

Copies the EDATemplate template to the working directory.
}
