% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windsorize.R
\name{windsorize}
\alias{windsorize}
\title{Windsorize}
\usage{
windsorize(Data, sdlim = 2.5)
}
\arguments{
\item{Data}{A numeric vector to be windsorized.}

\item{sdlim}{The number of standard deviations to use for the windsorization threshold. Values below
(mean - sdlim * sd) or above (mean + sdlim * sd) will be replaced with the threshold values.}
}
\value{
The windsorized numeric vector.
}
\description{
This function performs windsorization on a numeric vector, replacing values below or above a certain threshold
with the corresponding threshold value.
}
